consolidated word api
